+++ CHANGELOG.md
@@ -9,6 +9,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+- Add support for adding description to schema https://github.com/webonyx/graphql-php/pull/1834
+
 ## v15.29.4
 
 ### Fixed